Fixed the last typescript errors and divided the components that were nested in the tasks.ts (the dialogs)
