+++ FRONTEND/newFrontEnd/apTime/src/app/components/create-task-dialog/create-task-dialog.component.css
@@ -1,0 +1,14 @@
+.example-spacer {
+    flex: 1 1 auto;
+}
+
+.custom-button{
+    right: 20px;
+    bottom: 10px;
+    left: auto;
+    position:fixed !important;
+}
+
+.label-color {
+    color:#673AB7 !important;
+}
