+++ FRONTEND/newFrontEnd/apTime/src/app/components/create-task-dialog/create-task-dialog.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CreateTaskDialogComponent } from './create-task-dialog.component';
+
+describe('CreateTaskDialogComponent', () => {
+  let component: CreateTaskDialogComponent;
+  let fixture: ComponentFixture<CreateTaskDialogComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ CreateTaskDialogComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CreateTaskDialogComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
