+++ FRONTEND/newFrontEnd/apTime/src/app/components/delete-task-dialog/delete-task-dialog.component.html
@@ -1,0 +1,8 @@
+<h1 mat-dialog-title>Task "{{name}}"</h1>
+<div mat-dialog-content>
+    <p>Are you sure you want to delete the task?</p>
+</div>
+<div mat-dialog-actions>
+    <button mat-button (click)="onNoClick()">No</button>
+    <button mat-button (click)="onYesClick()" cdkFocusInitial>Yes</button>
+</div>
