+++ FRONTEND/newFrontEnd/apTime/src/app/components/delete-task-dialog/delete-task-dialog.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DeleteTaskDialogComponent } from './delete-task-dialog.component';
+
+describe('DeleteTaskDialogComponent', () => {
+  let component: DeleteTaskDialogComponent;
+  let fixture: ComponentFixture<DeleteTaskDialogComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ DeleteTaskDialogComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DeleteTaskDialogComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
