--- FRONTEND/newFrontEnd/apTime/src/app/components/tasks/confirm-delete-dialog.html
@@ -1,8 +1,0 @@
-<h1 mat-dialog-title>Task "{{name}}"</h1>
-<div mat-dialog-content>
-  <p>Are you sure you want to delete the task?</p>
-</div>
-<div mat-dialog-actions>
-  <button mat-button (click)="onNoClick()">No</button>
-  <button mat-button (click)="onYesClick()" cdkFocusInitial>Yes</button>
-</div>
