--- FRONTEND/newFrontEnd/apTime/src/app/components/tasks/edit-task-dialog.html
@@ -1,109 +1,0 @@
-
-<mat-toolbar color="primary" style="width:100%;">
-  <mat-toolbar-row>
-    <h1 mat-dialog-title>Edit Task "{{task.name}}"</h1>
-	<span class="example-spacer"></span>
-	<button mat-button  (click)="onNoClick()">
-		<mat-icon class="example-icon" aria-hidden="false" aria-label="Example user verified icon">close</mat-icon>
-	</button>
-</mat-toolbar-row>
-</mat-toolbar>
-<div mat-dialog-content>
-<form>
-  <div fxLayout="column">
-	  <mat-form-field>
-	    <mat-label>Name</mat-label>
-	    <input matInput [(ngModel)]="task.name" name="taskName">
-	  </mat-form-field>
-  
-
-  <mat-form-field>
-    <mat-label>Category</mat-label>
-    <input matInput [(ngModel)]="task.category" name="taskCat">
-  </mat-form-field>
-
-  <mat-form-field>
-    <mat-label>Description</mat-label>
-    <input matInput [(ngModel)]="task.description" name="taskDesc">
-  </mat-form-field>
-  
-  </div>
-
-
-  <div fxLayout="row" fxLayoutGap="15px">
-  	<div fxFlex="">
-	  <mat-form-field>
-		  <mat-label>Start date</mat-label>
-		  <input matInput [matDatepicker]="picker1" [(ngModel)]="task.scheduledstart" name="taskSDate">
-		  <mat-datepicker-toggle matSuffix [for]="picker1"></mat-datepicker-toggle>
-		  <mat-datepicker #picker1></mat-datepicker>
-		</mat-form-field>
-	</div>
-	<div fxFlex="" fxLayoutGap="15px">
-	  <mat-form-field>
-	    <mat-label>Start Time</mat-label>
-	    <input matInput [(ngModel)]="task.ssTime" name="taskssTime">
-	  </mat-form-field>
-	</div>
-  </div>
-
-  <div fxLayout="row"  fxLayoutGap="15px">
-  	<div fxFlex="">
-	  	<mat-form-field>
-		  <mat-label>End date</mat-label>
-		  <input matInput [matDatepicker]="picker2" [(ngModel)]="task.seDate" name="taskseDate">
-		  <mat-datepicker-toggle matSuffix [for]="picker2"></mat-datepicker-toggle>
-		  <mat-datepicker #picker2></mat-datepicker>
-		</mat-form-field>
-	</div>
-	<div fxFlex="" fxLayoutGap="15px">
-	  <mat-form-field>
-	    <mat-label>End Time</mat-label>
-	    <input matInput [(ngModel)]="task.seTime" name="taskseTime">
-	  </mat-form-field>
-	</div>
-  </div>
-  <button mat-button (click)="onActualsClick()" *ngIf="!showActuals"> Edit actuals <mat-icon class="example-icon" aria-hidden="false" aria-label="Edit actual times icon">edit</mat-icon></button>
-  
-  <div *ngIf="showActuals">
-  <mat-divider></mat-divider>
-  <p style="color:purple; !important">Actual Times</p>
-	<div fxLayout="row"  fxLayoutGap="15px">
-		<div fxFlex="">
-		<mat-form-field>
-			<mat-label>Start date</mat-label>
-			<input matInput [matDatepicker]="picker3" [(ngModel)]="task.asDate" name="taskasDate">
-			<mat-datepicker-toggle matSuffix [for]="picker3"></mat-datepicker-toggle>
-			<mat-datepicker #picker3></mat-datepicker>
-		  </mat-form-field>
-	  </div>
-	  <div fxFlex="" >
-		<mat-form-field>
-		  <mat-label>Start Time</mat-label>
-		  <input matInput [(ngModel)]="task.asTime" name="taskasTime">
-		</mat-form-field>
-	  </div>
-	</div>
-  
-	<div fxLayout="row"  fxLayoutGap="15px">
-		<div fxFlex="" >
-			<mat-form-field>
-			<mat-label>End date</mat-label>
-			<input matInput [matDatepicker]="picker4" [(ngModel)]="task.aeDate" name="taskaeDate">
-			<mat-datepicker-toggle matSuffix [for]="picker4"></mat-datepicker-toggle>
-			<mat-datepicker #picker4></mat-datepicker>
-		  </mat-form-field>
-	  </div>
-	  <div fxFlex="">
-		<mat-form-field>
-		  <mat-label>End Time</mat-label>
-		  <input matInput [(ngModel)]="task.aeTime" name="taskaeTime">
-		</mat-form-field>
-	  </div>
-	</div>
-</div>
-</form>
-</div>
-<div mat-dialog-actions align="end" class="header-text">
-  <button mat-raised-button color="primary" [mat-dialog-close]="data"  (click)="onYesClick()" cdkFocusInitial>Save</button>
-</div>
