+++ FRONTEND/newFrontEnd/apTime/src/app/components/tasks/tasks.component.ts
@@ -4,6 +4,9 @@ import { HttpClient } from '@angular/common/http';
 import { SessionService } from '../../services/session.service';
 import { HttpHeaders } from '@angular/common/http';
 import { Validators, FormBuilder, FormGroup, FormControl } from '@angular/forms';
+import { CreateTaskDialogComponent } from '../create-task-dialog/create-task-dialog.component';
+import { DeleteTaskDialogComponent } from '../delete-task-dialog/delete-task-dialog.component';
+import { EditTaskDialogComponent } from '../edit-task-dialog/edit-task-dialog.component';
 
 export interface DialogData {
   name: string;
@@ -44,7 +47,7 @@ export class TasksComponent implements OnInit {
 
   openDialog(): void {
     const dialogRef = this.dialog.open(
-    CreateTaskDialog, {
+    CreateTaskDialogComponent, {
       // width: '250px',
       data: {
       token: this.token
@@ -60,7 +63,7 @@ export class TasksComponent implements OnInit {
   openDeleteDialog(i: number, name: string): void {
 
     const dialogRef = this.dialog.open(
-    ConfirmDeleteDialog, {
+        DeleteTaskDialogComponent, {
       // width: '250px',
       data: {
       id: i,
@@ -76,7 +79,7 @@ export class TasksComponent implements OnInit {
 
   openEditDialog(i: number, index: number): void {
     const dialogRef = this.dialog.open(
-    EditTaskDialog, {
+    EditTaskDialogComponent, {
       data: {
       task:  this.tasks[index],
       token: this.token
@@ -101,234 +104,5 @@ export class TasksComponent implements OnInit {
       });
   }
 
-
-}
-
-@Component({
-  selector: 'create-task-dialog',
-  templateUrl: 'create-task-dialog.html',
-  styleUrls: ['./tasks.component.css']
-})
-export class CreateTaskDialog {
-
-  scheduledStart;
-  scheduledEnd;
-  timeRegex = /^(?:(?:1[0-2]|0?[1-9]):[0-5]\d\s*[AaPp][Mm])?$/;
-
-  nameFormControl = new FormControl('', [
-    Validators.required
-  ]);
-
-  sDateFormControl = new FormControl('', [
-    Validators.required
-  ]);
-
-  eDateFormControl = new FormControl('', [
-    Validators.required
-  ]);
-
-  sTimeFormControl = new FormControl('', [
-    Validators.required,
-    Validators.pattern(this.timeRegex)
-  ]);
-
-  eTimeFormControl = new FormControl('', [
-    Validators.required,
-    Validators.pattern(this.timeRegex)
-  ]);
-
-  constructor(
-    public dialogRef: MatDialogRef<CreateTaskDialog>,
-    private http: HttpClient,
-    private sessionService: SessionService,
-    @Inject(MAT_DIALOG_DATA) public data: DialogData) {}
-
-  onNoClick(): void {
-    this.dialogRef.close();
-  }
-
-  onYesClick(): void {
-
-    this.scheduledStart = this.dateConversion(this.data.ssTime, this.data.ssDate);
-    this.scheduledEnd = this.dateConversion(this.data.seTime, this.data.seDate);
-
-    console.log(this.scheduledStart);
-    console.log(this.scheduledEnd);
-    console.log('Before POST');
-
-    const headers = { Authorization: 'Bearer ' + this.sessionService.getToken()};
-
-    const body = { name: this.data.name,
-    description: this.data.description,
-    category: this.data.category,
-    scheduledstart: this.scheduledStart,
-    scheduledend: this.scheduledEnd,
-     };
-
-    this.http.post('http://localhost:8001/tasks/newtask', body, { headers }).subscribe({
-    next: data => console.log(data),
-    error: error => console.error('There was an error!', error)
-    });
-
-
-    this.dialogRef.close();
-  }
-
-  dateConversion(time: string, date: Date): Date {
-    let tempDate = date;
-    let tempTime = time;
-    console.log('IN dateconversion');
-    let parts = tempTime.match(/(\d+):(\d+) (AM|PM)/);
-    if (parts) {
-        let hours = parseInt(parts[1]),
-            minutes = parseInt(parts[2]),
-            tt = parts[3];
-        if (tt === 'PM' && hours < 12) { hours += 12; }
-        date.setHours(hours, minutes, 0, 0);
-    }
-
-    return tempDate;
-  }
-
-  ngOnInit() {
-    console.log(this.data.token);
-  }
-
-}
-
-@Component({
-  selector: 'confirm-delete-dialog',
-  templateUrl: 'confirm-delete-dialog.html',
-  styleUrls: ['./tasks.component.css']
-})
-export class ConfirmDeleteDialog {
-  id;
-  number;
-  name;
-
-
-  constructor(
-    public dialogRef: MatDialogRef<ConfirmDeleteDialog>,
-    private http: HttpClient,
-    private sessionService: SessionService,
-    @Inject(MAT_DIALOG_DATA) public data: DialogData) {data: {
-    }}
-
-  onNoClick(): void {
-    this.dialogRef.close();
-  }
-
-  onYesClick(): void {
-
-    const headers = { Authorization: 'Bearer ' + this.sessionService.getToken()};
-
-    this.http.delete('http://localhost:8001/tasks/task/' + this.data.id, { headers }).subscribe({
-    next: data => console.log(data),
-    error: error => console.error('There was an error!', error)
-    });
-
-    this.dialogRef.close();
-  }
-
-  ngOnInit() {
-    this.name = this.data.name;
-  }
-
-}
-
-
-@Component({
-  selector: 'edit-task-dialog',
-  templateUrl: 'edit-task-dialog.html',
-  styleUrls: ['./tasks.component.css']
-})
-export class EditTaskDialog {
-
-    id: number;
-    task: any;
-    showActuals: boolean;
-
-    scheduledStart;
-    scheduledEnd;
-    actualStart;
-    actualEnd;
-
-
-  constructor(
-    public dialogRef: MatDialogRef<EditTaskDialog>,
-    private http: HttpClient,
-    private sessionService: SessionService,
-    @Inject(MAT_DIALOG_DATA) public data: DialogData) {data: {
-
-    }}
-
-  onNoClick(): void {
-
-    this.dialogRef.close();
-  }
-
-  onYesClick(): void {
-
-    this.scheduledStart = this.dateConversion(this.task.ssTime, this.task.ssDate);
-    this.scheduledEnd = this.dateConversion(this.task.seTime, this.task.seDate);
-    this.actualStart = this.dateConversion(this.task.asTime, this.task.asDate);
-    this.actualEnd = this.dateConversion(this.task.aeTime, this.task.aeDate);
-
-    const headers = { Authorization: 'Bearer ' + this.sessionService.getToken()};
-
-    const body = {
-    id: this.task.id,
-    name: this.task.name,
-    description: this.task.description,
-    category: this.task.category,
-    scheduledstart: this.scheduledStart,
-    scheduledend: this.scheduledEnd,
-    actualstart: this.actualStart,
-    actualend: this.actualEnd
-
-     };
-
-    console.log('BEFORE put');
-    this.http.put('http://localhost:8001/tasks/task', body, { headers }).subscribe({
-    next: data => console.log(data),
-    error: error => console.error('There was an error!', error)
-    });
-
-
-    this.dialogRef.close();
-  }
-
-  onActualsClick(): void {
-    this.showActuals = true;
-    console.log('Actual click');
-  }
-
-  dateConversion(time: string, date: Date): Date {
-    let tempDate = date;
-    let tempTime = time;
-
-    let parts = tempTime.match(/(\d+):(\d+) (AM|PM)/);
-    if (parts) {
-        let hours = parseInt(parts[1]),
-            minutes = parseInt(parts[2]),
-            tt = parts[3];
-        if (tt === 'PM' && hours < 12) { hours += 12; }
-        date.setHours(hours, minutes, 0, 0);
-    }
-
-    return tempDate;
-  }
-
-
-  ngOnInit() {
-    this.showActuals = false;
-    this.id = this.data.id;
-    this.task = this.data.task;
-
-    // this.task.scheduledstart = new Date(this.task.scheduledstart)
-
-    console.log('The task: ', this.task);
-
-  }
 
 }
